Delete all versions and update documentation
