--- src/services/codesGenerator/config.ts
@@ -1,12 +1,0 @@
-import { IConfig } from './models/interfaces'
-
-const config: IConfig = {
-  algorithm: 'secuential',
-  config: {
-    amount: 5,
-    digits: 5,
-    startWith: 1,
-  },
-}
-
-export default config
