--- src/services/codesGenerator/models/interfaces.ts
@@ -1,8 +1,0 @@
-export interface IAlgorithm {
-  getCodes(arg: any): string[]
-}
-
-export interface IConfig {
-  algorithm: string
-  config: any
-}
