+++ src/services/couponsGenerator/Algorithm.ts
