+++ src/services/couponsGenerator/Algorithms/index.ts
