+++ src/services/couponsGenerator/config.ts
@@ -1,10 +1,12 @@
 import { IConfig } from './models/interfaces'
 
 const config: IConfig = {
-  algorithm: 'numeric',
-  amount: 5,
-  digits: 5,
-  startWith: 8673,
+  algorithm: 'secuential',
+  config: {
+    amount: 5,
+    digits: 5,
+    startWith: 1,
+  },
 }
 
 export default config
