+++ src/services/couponsGenerator/models/interfaces.ts
@@ -1,12 +1,8 @@
-export interface IAlgorithms {
-  secuential(initValue: number, digits: number, amount: number): string[]
-  alphanumeric(digits: number, amount: number, pattern: string): string[]
+export interface IAlgorithm {
+  getCodes(arg: any): string[]
 }
 
 export interface IConfig {
   algorithm: string
-  amount?: number
-  digits?: number
-  startWith?: number
-  pattern?: string
+  config: any
 }
